Adds basic docs explaining abstractions.
readthedocs.org is a convenient site for hosting documentation,
with the added benefit that Mozilla helps host the site.
This adds three short pages of docs that go over some of
the main concepts I've had to learn about since starting
looking at Gladius. It uses Sphinx for building the docs
as readthedocs.org uses Sphinx.