Remove jake imports from Jakefile and tasks.
Having `var jake = require('jake')` in a Jakefile or an included
task results in the script looking for a locally-installed version
of jake, when jake is supposed to be installed globally. And even
if you have a local install, you have to explicitly run it to get
the task to run.

But, unless I'm missing something, you can just remove the jake
imports and everything will work (as jake is already included
if you're running the jake file). Magic!